feat: trip-event-presenter
