+++ src/presenter/events-presenter.js
@@ -1,8 +1,7 @@
-import {render, replace} from '../framework/render.js';
+import {render} from '../framework/render.js';
 import SortView from '../view/sort.js';
 import TripEventsListView from '../view/trip-events-list.js';
-import TripEventsItemView from '../view/trip-events-item.js';
-import EditPointView from '../view/edit-point.js';
+import tripEventPresenter from '../presenter/trip-event-presenter.js';
 
 export default class EventsPresenter {
   #sortComponent = new SortView();
@@ -13,10 +12,6 @@ export default class EventsPresenter {
   #destinationModel = null;
   #offersModel = null;
   #eventPoints = [];
-  #editPoint = null;
-  #destination = null;
-  #offer = null;
-  #offers = null;
 
   constructor({container, eventPointsModel, editPointModel, destinationModel, offersModel}) {
     this.#container = container;
@@ -28,54 +23,18 @@ export default class EventsPresenter {
 
   init() {
     this.#eventPoints = [...this.#eventPointsModel.get()];
-    this.#editPoint = this.#editPointModel.get()[0];
-    this.#destination = this.#destinationModel.getById(this.#editPoint.destination);
-    this.#offer = this.#offersModel.getByType(this.#editPoint.type);
-    this.#offers = this.#offer.offers;
 
     render(this.#sortComponent, this.#container);
     render(this.#tripEventsListComponent, this.#container);
 
     for (let i = 0; i < this.#eventPoints.length; i++) {
-      this.#renderPoints(
-        this.#eventPoints[i],
-        this.#editPoint,
-        this.#offers,
-        this.#destination
-      );
+      new tripEventPresenter({
+        container: this.#tripEventsListComponent.element,
+        editPointModel: this.#editPointModel,
+        destinationModel: this.#destinationModel,
+        offersModel: this.#offersModel,
+        eventPoint: this.#eventPoints[i],
+      }).init();
     }
-  }
-
-  #renderPoints(points, editPoint, offers, destination) {
-    const pointComponent = new TripEventsItemView({
-      points,
-      onClick: rollupBtnClick
-    });
-
-    const editComponent = new EditPointView({
-      editPoint,
-      offers,
-      destination,
-      onSubmit: closeEditOpenPoint
-    });
-
-    const escKeyEventEdit = (evt) => {
-      if (evt.key === 'Escape') {
-        evt.preventDefault();
-        closeEditOpenPoint();
-        document.removeEventListener('keydown', escKeyEventEdit);
-      }
-    };
-
-    function rollupBtnClick() {
-      replace(editComponent, pointComponent);
-      document.addEventListener('keydown', escKeyEventEdit);
-    }
-
-    function closeEditOpenPoint() {
-      replace(pointComponent, editComponent);
-    }
-
-    render(pointComponent, this.#tripEventsListComponent.element);
   }
 }
